feat : 현재 page 관련 정보를 data.js에 저장
- page 관련 정보를 object 형태로 data.js에 저장
- changeClickNavTab에서 currentPage 정보를 인자로 받아 함수의 범용성을 높힘
