+++ public/javascripts/data.js
@@ -1,0 +1,3 @@
+const headerNavData = { currentPage: 0 };
+
+export { headerNavData };
