+++ views/index.html
@@ -9,6 +9,7 @@
   </head>
   <body>
     <script type="module" src="javascripts/utils.js"></script>
+    <script type="module" src="javascripts/data.js"></script>
     <script type="module" src="javascripts/genre-nav.js"></script>
     <script type="module" src="javascripts/main.js"></script>
     <div class="header">
